fix(error)!: Move ErrorConvert to error module
